四期直播/19.爬虫&APP逆向/16.B站app注册/geetest/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
pyasn1==0.4.8
pycryptodome==3.15.0
PyExecJS==1.5.1
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
